+++ src/tg2go/bot/handlers/middleware.py
@@ -5,6 +5,8 @@ from aiogram import BaseMiddleware, Dispatcher
 from aiogram.filters.callback_data import CallbackData
 from aiogram.types import CallbackQuery, Message
 
+from tg2go.bot.handlers.staff.menus.common import staff_menu
+from tg2go.bot.handlers.staff.menus.panel import PanelMenu
 from tg2go.bot.lib.chat.block import CheckIfBlocked
 from tg2go.bot.lib.message.io import (
     ReceiveCallback,
@@ -14,7 +16,7 @@ from tg2go.bot.lib.message.io import (
 from tg2go.bot.lifecycle.active import bot_state
 
 
-class MessageLoggingMiddleware(BaseMiddleware):
+class LoggingMessageMiddleware(BaseMiddleware):
     async def __call__(
         self,
         handler: Callable[[Message, dict[str, Any]], Awaitable[Any]],
@@ -26,6 +28,16 @@ class MessageLoggingMiddleware(BaseMiddleware):
 
         await ReceiveMessage(message)
 
+        return await handler(message, data)
+
+
+class ClientMessageMiddleware(BaseMiddleware):
+    async def __call__(
+        self,
+        handler: Callable[[Message, dict[str, Any]], Awaitable[Any]],
+        message: Message,  # type: ignore[override]
+        data: dict[str, Any],
+    ) -> Any:
         if not bot_state.active and message.text == "/start":
             await SendMessage(
                 chat_id=message.chat.id,
@@ -36,7 +48,7 @@ class MessageLoggingMiddleware(BaseMiddleware):
         return await handler(message, data)
 
 
-class CallbackLoggingMiddleware(BaseMiddleware):
+class LoggingCallbackMiddleware(BaseMiddleware):
     async def __call__(
         self,
         handler: Callable[[CallbackQuery, dict[str, Any]], Awaitable[Any]],
@@ -46,6 +58,7 @@ class CallbackLoggingMiddleware(BaseMiddleware):
         assert isinstance(callback_query.message, Message)
 
         if await CheckIfBlocked(callback_query.message.chat.id):
+            await callback_query.answer()
             return
 
         callback_data = data.get("callback_data")
@@ -56,16 +69,63 @@ class CallbackLoggingMiddleware(BaseMiddleware):
             data=callback_data,
         )
 
+        return await handler(callback_query, data)
+
+
+class ClientCallbackMiddleware(BaseMiddleware):
+    async def __call__(
+        self,
+        handler: Callable[[CallbackQuery, dict[str, Any]], Awaitable[Any]],
+        callback_query: CallbackQuery,  # type: ignore[override]
+        data: dict[str, Any],
+    ) -> Any:
+        assert isinstance(callback_query.message, Message)
+
+        callback_data = data.get("callback_data")
+        assert isinstance(callback_data, CallbackData)
+
         if not bot_state.active and "client" in callback_data.__prefix__:
             await SendMessage(
                 chat_id=callback_query.message.chat.id,
                 text="Бот не активен.",
             )
+            await callback_query.answer()
+            return
+
+        return await handler(callback_query, data)
+
+
+class StaffCallbackMiddleware(BaseMiddleware):
+    async def __call__(
+        self,
+        handler: Callable[[CallbackQuery, dict[str, Any]], Awaitable[Any]],
+        callback_query: CallbackQuery,  # type: ignore[override]
+        data: dict[str, Any],
+    ) -> Any:
+        assert isinstance(callback_query.message, Message)
+
+        callback_data = data.get("callback_data")
+        assert isinstance(callback_data, CallbackData)
+
+        if (
+            bot_state.active
+            and "staff" in callback_data.__prefix__
+            and "staff.active" != callback_data.__prefix__
+        ):
+            await staff_menu.ChangeToNewMenu(
+                callback_query=callback_query,
+                new_menu=PanelMenu(),
+            )
+            await callback_query.answer()
             return
 
         return await handler(callback_query, data)
 
 
 def SetBotMiddleware(dp: Dispatcher) -> None:
-    dp.message.middleware(MessageLoggingMiddleware())
-    dp.callback_query.middleware(CallbackLoggingMiddleware())
+    dp.message.middleware(LoggingMessageMiddleware())
+    dp.message.middleware(ClientMessageMiddleware())
+
+    dp.callback_query.middleware(LoggingCallbackMiddleware())
+    dp.callback_query.middleware(ClientCallbackMiddleware())
+    dp.callback_query.middleware(StaffCallbackMiddleware())