+++ src/tg2go/bot/handlers/staff/callbacks/good/add.py
@@ -6,7 +6,7 @@ from aiogram.fsm.context import FSMContext
 from aiogram.fsm.state import State, StatesGroup
 
 from tg2go.bot.handlers.staff.menus.category.category import CategoryMenu
-from tg2go.bot.handlers.staff.menus.common import menu
+from tg2go.bot.handlers.staff.menus.common import staff_menu
 from tg2go.bot.lib.message.image import GetGoodImageDir, Image
 from tg2go.bot.lib.message.io import ContextIO, SendMessage
 from tg2go.services.staff.good import StaffGoodService
@@ -116,7 +116,7 @@ async def CommandStaffAddGoodImageUrl(
     )
     await state.clear()
 
-    await menu.SendMenu(
+    await staff_menu.SendMenu(
         chat_id=message.chat.id,
         menu=await CategoryMenu(data["category_id"]),
     )