+++ src/tg2go/bot/handlers/staff/callbacks/good/remove.py
@@ -1,7 +1,7 @@
 from aiogram import F, Router, types
 
 from tg2go.bot.handlers.staff.menus.category.category import CategoryMenu
-from tg2go.bot.handlers.staff.menus.common import menu
+from tg2go.bot.handlers.staff.menus.common import staff_menu
 from tg2go.bot.handlers.staff.menus.good.good import GoodMenu
 from tg2go.bot.handlers.staff.menus.good.remove import (
     GoodRemoveAction,
@@ -24,7 +24,7 @@ async def GoodRemoveDelete(
 
     good = await srv.GetGood(callback_data.good_id)
 
-    await menu.ChangeToNewMenu(
+    await staff_menu.ChangeToNewMenu(
         callback_query=callback_query,
         new_menu=await CategoryMenu(good.category_id),
     )
@@ -36,7 +36,7 @@ async def GoodRemoveBack(
     callback_query: types.CallbackQuery,
     callback_data: GoodRemoveCallbackData,
 ) -> None:
-    await menu.ChangeToNewMenu(
+    await staff_menu.ChangeToNewMenu(
         callback_query=callback_query,
         new_menu=await GoodMenu(callback_data.good_id),
     )