+++ src/tg2go/bot/handlers/staff/callbacks/panel.py
@@ -1,7 +1,9 @@
 from aiogram import F, Router, types
 
-from tg2go.bot.handlers.staff.menus.common import menu
+from tg2go.bot.handlers.staff.menus.common import staff_menu
 from tg2go.bot.handlers.staff.menus.panel import (
+    ActivePanelAction,
+    ActivePanelCallbackData,
     PanelAction,
     PanelCallbackData,
     PanelMenu,
@@ -16,18 +18,7 @@ router = Router()
 async def PanelActivate(callback_query: types.CallbackQuery) -> None:
     bot_state.Activate()
 
-    await menu.ChangeToNewMenu(
-        callback_query=callback_query,
-        new_menu=PanelMenu(),
-    )
-    await callback_query.answer()
-
-
-@router.callback_query(PanelCallbackData.filter(F.action == PanelAction.Deactivate))
-async def PanelDeactivate(callback_query: types.CallbackQuery) -> None:
-    bot_state.Deactivate()
-
-    await menu.ChangeToNewMenu(
+    await staff_menu.ChangeToNewMenu(
         callback_query=callback_query,
         new_menu=PanelMenu(),
     )
@@ -36,7 +27,7 @@ async def PanelDeactivate(callback_query: types.CallbackQuery) -> None:
 
 @router.callback_query(PanelCallbackData.filter(F.action == PanelAction.Settings))
 async def PanelSettings(callback_query: types.CallbackQuery) -> None:
-    await menu.ChangeToNewMenu(
+    await staff_menu.ChangeToNewMenu(
         callback_query=callback_query,
         new_menu=await SettingsMenu(),
     )
@@ -45,6 +36,29 @@ async def PanelSettings(callback_query: types.CallbackQuery) -> None:
 
 @router.callback_query(PanelCallbackData.filter(F.action == PanelAction.Exit))
 async def PanelExit(callback_query: types.CallbackQuery) -> None:
+    assert isinstance(callback_query.message, types.Message)
+
+    await callback_query.message.edit_reply_markup(reply_markup=None)
+    await callback_query.answer()
+
+
+@router.callback_query(
+    ActivePanelCallbackData.filter(F.action == ActivePanelAction.Deactivate)
+)
+async def ActivePanelDeactivate(callback_query: types.CallbackQuery) -> None:
+    bot_state.Deactivate()
+
+    await staff_menu.ChangeToNewMenu(
+        callback_query=callback_query,
+        new_menu=PanelMenu(),
+    )
+    await callback_query.answer()
+
+
+@router.callback_query(
+    ActivePanelCallbackData.filter(F.action == ActivePanelAction.Exit)
+)
+async def ActivePanelExit(callback_query: types.CallbackQuery) -> None:
     assert isinstance(callback_query.message, types.Message)
 
     await callback_query.message.edit_reply_markup(reply_markup=None)