+++ src/tg2go/db/models/good.py
@@ -74,11 +74,9 @@ class Good(Base, TimestampMixin):
         return f"<b>{self.name}, {self.price_rub}₽</b>\n\n{self.description}"
 
     def GetStaffInfo(self) -> str:
-        if self.valid:
-            valid = "Доступен для покупки"
+        if self.available:
+            available = "Позиция доступна для покупки"
         else:
-            valid = "<b>Недоступен для покупки</b>"
+            available = "❌ <b>Позиция недоступна для покупки</b>"
 
-        return f"<b>{self.name}, {self.price_rub}₽</b>\n\n{self.description}\n\nНаходится в категории: {self.category.name}.\n\n{valid}"
-
-        return f"- Категория позиции: {self.category.name}\n- Название: {self.name}\n- Цена: {self.price_rub}₽\n- Доступен для покупки: {'Да' if self.valid else 'Нет'}\n- Описание: {self.description}"
+        return f"<b>{self.name}, {self.price_rub}₽</b>\n\n{self.description}\n\nНаходится в категории: {self.category.name}.\n\n{available}."