+++ CHANGELOG.md
@@ -26,7 +26,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ---
 
-## [0.1.0] - 2025-06-27 - Vlad
+## [0.1.0] - 2025-06-30 - Vlad
+
+### Updated
+
+- Singlton staff menu adjustments
+- Middleware restructuring
+
+## [0.1.0] - 2025-06-28 - Vlad
 
 ### Updated